Migrating flags off main.go to a separate package
